Ajustes v-4

+++ DentistaApi/Models/Dashbords.cs
@@ -7,6 +7,6 @@
         public string[] dentistas { get; set; }
         public int[] qtdPorDentista { get; set; }
         public string[] espec {  get; set; }
-        public double[] qtdPorEspec { get; set; }
+        public int[] qtdPorEspec { get; set; }
     }
 }
